Backwards compatible with old p2p
