+++ salad/client/scripts/configure.py
@@ -4,14 +4,12 @@ This script uses the enigma provider class to identify all the network parameter
 and writes them to the salad client's `.env` file.
 """
 
-import os
 import socket
 from time import sleep
 
-import typing
-
 from enigma_docker_common.config import Config
 from enigma_docker_common.logger import get_logger
+from enigma_docker_common.utils import parse_env_file, dump_env_file
 
 logger = get_logger('salad-client-startup')
 
@@ -30,35 +28,12 @@ required = [  # required by provider AND locally
     'KEY_MANAGEMENT_DISCOVERY',
 ]
 
-env_defaults = {
-    'K8S': 'config/k8s_config.json',
-    'TESTNET': 'config/testnet_config.json',
-    'MAINNET': 'config/mainnet_config.json',
-    'COMPOSE': 'config/compose_config.json'
-}
-
-env = os.getenv('ENIGMA_ENV', 'COMPOSE')
 
 try:
-    config = Config(required=required, config_file=env_defaults[os.getenv('ENIGMA_ENV', 'COMPOSE')])
+    config = Config(required=required)
 except (ValueError, IOError) as e:
     logger.critical(f'encountered unexpected error while configuring the salad client: {e!r}')
     raise
-
-
-def parse_env_file(file: typing.Iterable[typing.Text]) -> dict:
-    """Parse a .env file to a dict"""
-    return dict(
-        # we ignore the types because mypy doesn't realize dict can take a Generator[List] where len(list) == 2.
-        line.rstrip().split('=', 2)  # type: ignore
-        for line
-        in file
-        if not line.startswith('#')
-    )
-
-
-def dump_env_file(env_vars: dict, file: typing.TextIO) -> None:
-    file.writelines(f'{key}={value}\n' for key, value in env_vars.items())
 
 
 def wait_for_operator_server():
