+++ salad/operator/scripts/configure.py
@@ -4,12 +4,10 @@ This script uses the enigma provider class to identify all the network parameter
 and writes them to the operator's `.env` file.
 """
 
-import os
-import typing
-
 from enigma_docker_common.config import Config
-from enigma_docker_common.provider import Provider
 from enigma_docker_common.logger import get_logger
+from enigma_docker_common.provider import Provider
+from enigma_docker_common.utils import parse_env_file, dump_env_file
 
 logger = get_logger('operator-startup')
 
@@ -17,44 +15,20 @@ logger = get_logger('operator-startup')
 required = [  # required by provider AND locally
     # defaults in local config file
     'CONTRACT_DISCOVERY_ADDRESS',
+    'MONGO_URL',
     'ENG_NODE_ADDRESS',
     'ENG_NODE_PORT',
     'ETH_NODE_PORT',
     'ETH_NODE_ADDRESS',
-    'MONGO_URL',
     'DB_NAME',
     'KEY_MANAGEMENT_DISCOVERY',
 ]
 
-env_defaults = {
-    'K8S': 'config/k8s_config.json',
-    'TESTNET': 'config/testnet_config.json',
-    'MAINNET': 'config/mainnet_config.json',
-    'COMPOSE': 'config/compose_config.json'
-}
-
-env = os.getenv('ENIGMA_ENV', 'COMPOSE')
-
 try:
-    config = Config(required=required, config_file=env_defaults[os.getenv('ENIGMA_ENV', 'COMPOSE')])
+    config = Config(required=required)
 except (ValueError, IOError) as e:
     logger.critical(f'encountered unexpected error while configuring the operator: {e!r}')
     raise
-
-
-def parse_env_file(file: typing.Iterable[typing.Text]) -> dict:
-    """Parse a .env file to a dict"""
-    return dict(
-        # we ignore the types because mypy doesn't realize dict can take a Generator[List] where len(list) == 2.
-        line.rstrip().split('=', 2)  # type: ignore
-        for line
-        in file
-        if not line.startswith('#')
-    )
-
-
-def dump_env_file(env_vars: dict, file: typing.TextIO) -> None:
-    file.writelines(f'{key}={value}\n' for key, value in env_vars.items() if value not in ['', [], (), None])
 
 
 def main():
@@ -66,7 +40,7 @@ def main():
     # parse, modify, and rewrite the .env files
     for path in ['.env', 'operator/.env']:
         with open(path, 'r') as file:
-            env_vars = parse_env_file(file)
+            envs = parse_env_file(file)
 
         for env_var, config_var in {
                 'ETH_HOST': 'ETH_NODE_ADDRESS',
@@ -79,13 +53,13 @@ def main():
                 'SALAD_SMART_CONTRACT_ADDRESS': 'SALAD_SMART_CONTRACT_ADDRESS',
                 'SALAD_SECRET_CONTRACT_ADDRESS': 'SALAD_SECRET_CONTRACT_ADDRESS',
         }.items():
-            env_vars[env_var] = config[config_var]
+            envs[env_var] = config[config_var]
 
-        env_vars['ENIGMA_CONTRACT_ADDRESS'] = enigma_contract_address
-        env_vars['ENIGMA_TOKEN_CONTRACT_ADDRESS'] = enigma_token_contract_address
+        envs['ENIGMA_CONTRACT_ADDRESS'] = enigma_contract_address
+        envs['ENIGMA_TOKEN_CONTRACT_ADDRESS'] = enigma_token_contract_address
 
         with open(path, 'w') as file:
-            dump_env_file(env_vars, file)
+            dump_env_file(envs, file)
 
 
 if __name__ == '__main__':
